requirements.txt
grpcio
grpcio-tools==1.12.0
mock
tox
testtools
pytest
setuptools
pbr>=2.0.0